Docs: Get documentation version from VERSION file
The version of the documentation is configured in the conf.py file. The
version has not been increased there for a longer time. On Read the
Docs you can find the version information for the library in the title
for example. So it looks like the documentation refers to an older
version of the library.

The current version of the library is in the atlassian/VERSION file. In
oder to keep the process simple, the documentation now also takes the
version information from this file. By that the version information of
the library and the documentation is always in sync.