All: Fixes typos (#673)
* All: Fixes typos

Scanned over whole repository and fixes some typos.

* All: Correction of typos in methods

Renaming methods can be dangerous for software that is already running
with older versions of the library. Therefore the renaming should be
done softly. Old methods are marked as deprecated and a fixed method is
created.

In the confluence script there was already a refactored function. So
this typo change was reverted.

In the bitbucket script a new function with correct method name is
introduced. The typo method was set to deprecated.