[RestClient] Added advanced_mode to POST, PUT, DELETE
Added boolean switch to post, put and delete methods of RestClient, to
make them align with the get implementation. This allows the caller to
retrieve raw response for individual calls rather than change the
property of the RestClient instance.

Added tests that mimic Confluence module behavior that check the type of
response

Closes atlassian-api/atlassian-python-api#681