Bitbucket Cloud: Set default url
Unlike the server variants, the url for the Bitbucket cloud is always
the same. There is no need to write the url every time, when the Cloud
class should be used. Also for new users of this library it can be
confusing, where to find the Bitbucket cloud url.

Therefore, this commit sets for the url parameter of the Bitbucket
Cloud class to a default value. It still can be modified by giving
another url as parameter (i.e. for tests)