[ISSUE-1242]  prevented weird exceptions related to integer handling (#1244)
if keys or values in the response json are `int`s, the `k + ": " + v`
will fail because of an `int` is not a `str`.  In order to avoid this,
we can either use `%`-formatting, `.format` formatting, or f-strings.
We chose `.format` for compatability, but regardless of which direction, the
issue should be addressed one of those ways.