[Jira] Fix get method data retrieval failures when advanced mode is on (#1449)
* [Rest Client] Add helper method to get content

_get_response_content is to be used for chaining together response
retrieval. This is useful in the case where you are querying the API and
then immediately extracting some data from the response. This method is
needed to account for the case where advanced mode is on and the initial
get call returns a requests.Response rather than a dictionary.

* [Jira] Update retrieval method calls

For methods that query the API and then immediately retrieve some data
from the response, replace the get call with the new
_get_response_content method to account for advanced mode usage.

* [Jira] Fix get_issue_labels data retrieval