[Jira] Update Agile (Greenhopper) REST API URL handling
This PR fixes the issue of creating URLs for resources from the Agile (Greenhopper) REST API group. The resource URLs were built incorrectly if we used the OAuth2 authorization method when creating a Jira client object.

The problem was that everything worked fine when auth, for example, by username/user_api_token, because, in this case, we could use the “api_root” default value from the Jira client constructor. However, when we use OAuth2, the "api_root" parameter must be overridden, but previously, before the fix, this override did not affect the composition of URLs to the above resources.
So, this PR fixes this issue with minimal related reformatting, which makes the code of the corresponding methods more consistent.